comercio exterior Requerimiento de exportacion

+++ app/src/main/java/com/ssl/jv/gip/util/EstadoRequerimientoExportacion.java
@@ -4,19 +4,19 @@ public enum EstadoRequerimientoExportacion {
 	
 	Nodiligenciado(0, "No diligenciado"), EsperandoAprobacion(1, "Esperando Aprobacion"), Aprobado(2,"Aprobado");
 
-		  private long codigo;
+		  private Integer codigo;
 		  private String nombre;
 
-		  private EstadoRequerimientoExportacion(long codigo) {
+		  private EstadoRequerimientoExportacion(Integer codigo) {
 		    this.codigo = codigo;
 		  }
 
-		  private EstadoRequerimientoExportacion(long codigo, String nombre) {
+		  private EstadoRequerimientoExportacion(Integer codigo, String nombre) {
 		    this.codigo = codigo;
 		    this.nombre = nombre;
 		  }
 
-		  public long getCodigo() {
+		  public Integer getCodigo() {
 		    return codigo;
 		  }
